first stamina drain implementation Bugs: 1. In D2 when Stamina is drained it does not regenerate until you completely stop 2. Stamina Bar does not get red when near the end

+++ d2core/d2map/d2mapentity/factory.go
@@ -77,7 +77,7 @@ func (f *MapEntityFactory) NewPlayer(id, name string, x, y, direction int, heroT
 	}
 
 	stats.NextLevelExp = f.asset.Records.GetExperienceBreakpoint(heroType, stats.Level)
-	stats.Stamina = stats.MaxStamina
+	stats.Stamina = float64(stats.MaxStamina)
 
 	defaultCharStats := f.asset.Records.Character.Stats[heroType]
 	statsState := f.HeroStateFactory.CreateHeroStatsState(heroType, defaultCharStats)
@@ -249,7 +249,7 @@ func (f *MapEntityFactory) NewCastOverlay(x, y int, overlayRecord *d2records.Ove
 	result := &CastOverlay{
 		AnimatedEntity: entity,
 		record:         overlayRecord,
-		playLoop: playLoop,
+		playLoop:       playLoop,
 	}
 
 	return result, nil
